% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akNN.R
\name{FindNeighbors_akNN}
\alias{FindNeighbors_akNN}
\title{akNN}
\usage{
FindNeighbors_akNN(
  obj,
  reduction = "pca",
  knn = 20,
  prune = 1/15,
  delta = -0.5,
  dims = 50
)
}
\arguments{
\item{obj}{An Seurat object.}

\item{reduction}{Name of reduction to pull cell embeddings for. Default is 'pca'}

\item{knn}{Defines the largest k. Default is 20}

\item{prune}{The cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Default is 1/15}

\item{delta}{Default is -0.5}

\item{dims}{Dimensions of reduction to use. Default is 50}
}
\value{
This function will return a graph object name as akNN.
}
\description{
An Adjusted K-Nearst-Neighbors Graph Construction Method for the Clustering of Single Cell RNA Sequencing Data
}
\examples{
library(akNN)
load(system.file("data", "humanPancreas_GSE84133.Rdata", package = "akNN"))
pancreas <- FindNeighbors_akNN(pancreas)
dim(pancreas@graphs$akNN)
}
